[LiveCleanup] Allow some sort of file composition for Live Auction files
