+++ Artsy/View_Controllers/Live_Auctions/ViewModels/LiveAuctionEventViewModel.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+class LiveAuctionEventViewModel : NSObject {
+    let event: LiveEvent
+
+    init(event:LiveEvent) {
+        self.event = event
+    }
+}
